Mod 10 calculator
Calculates a check sum to compare to the last digit of a number as
follows:
take every other digit (1's, 100's, ect.) and multiply the number by 2.
add all these sums, plus the individual digits leftover.
the modulus 10 of the final sum is the 'check sum'.  this number will be
appended to the right side of the number as so:
original number 12345
1_3_5 brought down and multiplied by 2
2, 6, 10 result.
add remaining numbers to get
2, 2, 6, 4, 10
add all numbers to get
24
24%10 = 6
final 'test code' would be 123456.